organise actions & reducers with store in to folders
