+++ src/components/auth/Login.js
@@ -3,7 +3,7 @@ import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
-import { authenticateUser, clearStatus } from '../../action';
+import { authenticateUser, clearStatus } from '../../actions/action';
 import LoginWrapper from './loginStyle';
 import image from '../../asset/image/eventCenter.jpg';
