+++ src/components/details/detailsPage.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import { sheduleMeeting } from '../../action';
+import { sheduleMeeting } from '../../actions/action';
 import DetailWrapper from './detailStyle';
 
 const DetailPage = ({ centers, bookings, match }) => {
